+++ app/Http/Controllers/PermissionController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Permission;
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class PermissionController extends Controller
 {
@@ -15,7 +16,8 @@ class PermissionController extends Controller
      */
     public function index()
     {
-        $permission= Permission::orderBy('id','Asc')->paginate(5);
+        Gate::authorize('haveaccess', 'menu.index');
+        $permission= Permission::orderBy('id','Asc')->paginate(7);
     
         return view('Permissions.indexper',['permissions'=>$permission]);
     }
@@ -27,6 +29,7 @@ class PermissionController extends Controller
      */
     public function create()
     {
+        Gate::authorize('haveaccess', 'menu.create');
         return \view('Permissions.createper');
     }
 
@@ -38,6 +41,7 @@ class PermissionController extends Controller
      */
     public function store(Request $request)
     {
+        Gate::authorize('haveaccess', 'menu.store');
         $request->validate([
 
             'namep'      => 'required|string|max:150',
@@ -61,6 +65,7 @@ class PermissionController extends Controller
      */
     public function show(Permission $permission)
     {
+        Gate::authorize('haveaccess', 'menu.show');
         return \view ('Permissions.showper',['permisos' =>$permission]);
     }
 
@@ -72,6 +77,7 @@ class PermissionController extends Controller
      */
     public function edit(Permission $permission)
     {
+        Gate::authorize('haveaccess', 'menu.edit');
         return \view('Permissions.editper',['permission'=>$permission]);
     }
 
@@ -84,6 +90,7 @@ class PermissionController extends Controller
      */
     public function update(Request $request, Permission $permission)
     {
+        Gate::authorize('haveaccess', 'menu.update');
         $request->validate([
 
             'namep'      => 'required|string|max:150',
@@ -106,6 +113,7 @@ class PermissionController extends Controller
      */
     public function destroy($id)
     {
+        Gate::authorize('haveaccess', 'menu.destroy');
         $permission= Permission::find($id);
         $permission->delete();
